+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added seekability to ShareFileClient.OpenRead().
 - Renamed ShareClient.SetTier() -> ShareClient.SetProperties().  SetProperties() can be used to set both Share Tier and Share Quota.
 - Changed ShareDeleteOptions.IncludeSnapshots -> .ShareSnapshotsDeleteOption, and added option to also delete Share Snapshots that have been leased.
+- Added additional info to exception messages.
 
 ## 12.5.0-preview.1 (2020-09-30)
 - Added support for service version 2020-02-10.